Run `invoke static` while compiling the source.
Previously, the `invoke static` command was executed while starting systemd
inventree-server job. However, this caused a problem as the /nix/store is
read-only and for some reasons `invoke prerender` executed by `invoke static`
requires the static_root to not be static.

Thus, the `invoke static` is moved to the source bundling operation, and copied
over to the static_root directory, as apparently the static root directory is
also used for storing some mutable files.